+++ jdaviz/configs/imviz/plugins/__init__.py
@@ -1,4 +1,4 @@
 from .tools import *  # noqa
 from .viewers import *  # noqa
 from .image_viewer_creator import *  # noqa
-from .coordinates_overlay import *  # noqa
+from .coords_info import *  # noqa
