--- jdaviz/configs/imviz/plugins/coordinates_overlay/__init__.py
@@ -1,1 +1,0 @@
-from .coordinates_overlay import *
