+++ jdaviz/configs/imviz/plugins/coords_info/__init__.py
@@ -1,0 +1,1 @@
+from .coords_info import *
