+++ jdaviz/configs/imviz/plugins/coords_info/coords_info.py
@@ -9,5 +9,5 @@ __all__ = ['CoordsInfo']
 
 @info_registry('g-coords-info')
 class CoordsInfo(TemplateMixin):
-    template = load_template("coordinates_overlay.vue", __file__).tag(sync=True)
+    template = load_template("coords_info.vue", __file__).tag(sync=True)
     text = Unicode("").tag(sync=True)
