+++ jdaviz/configs/imviz/plugins/coords_info/coords_info.vue
